backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.20
opencv-python==4.12.0.88
pillow==11.3.0
requests==2.32.4
elevenlabs==2.8.1
transformers==4.54.1
datasets==4.0.0
nltk==3.9.1
torch==2.7.1
accelerate==1.2.1
moviepy==2.2.1
pyjwt==2.9.0
passlib[bcrypt]==1.7.4
cryptography==43.0.1
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1
pytest-mock==3.14.0
pytest-cov==6.0.0
prometheus-client==0.20.0
prometheus-fastapi-instrumentator==6.1.0